add component for shape indicators
