+++ packages/editor/src/lib/components/LiveCollaborators.tsx
@@ -9,13 +9,7 @@ import { usePresence } from '../hooks/usePresence'
 
 export const LiveCollaborators = track(function Collaborators() {
 	const peerIds = usePeerIds()
-	return (
-		<>
-			{peerIds.map((id) => (
-				<CollaboratorGuard key={id} collaboratorId={id} />
-			))}
-		</>
-	)
+	return peerIds.map((id) => <CollaboratorGuard key={id} collaboratorId={id} />)
 })
 
 const CollaboratorGuard = track(function CollaboratorGuard({
