+++ packages/editor/src/lib/components/default-components/DefaultCanvas.tsx
@@ -33,7 +33,7 @@ export interface TLCanvasComponentProps {
 export function DefaultCanvas({ className }: TLCanvasComponentProps) {
 	const editor = useEditor()
 
-	const { Background, SvgDefs } = useEditorComponents()
+	const { Background, SvgDefs, ShapeIndicators } = useEditorComponents()
 
 	const rCanvas = useRef<HTMLDivElement>(null)
 	const rHtmlLayer = useRef<HTMLDivElement>(null)
@@ -161,7 +161,7 @@ export function DefaultCanvas({ className }: TLCanvasComponentProps) {
 						<BrushWrapper />
 						<ScribbleWrapper />
 						<ZoomBrushWrapper />
-						<ShapeIndicators />
+						{ShapeIndicators && <ShapeIndicators />}
 						<HintedShapeIndicator />
 						<SnapIndicatorWrapper />
 						<SelectionForegroundWrapper />
@@ -201,18 +201,9 @@ function ScribbleWrapper() {
 
 	if (!(Scribble && scribbles.length)) return null
 
-	return (
-		<>
-			{scribbles.map((scribble) => (
-				<Scribble
-					key={scribble.id}
-					className="tl-user-scribble"
-					scribble={scribble}
-					zoom={zoomLevel}
-				/>
-			))}
-		</>
-	)
+	return scribbles.map((scribble) => (
+		<Scribble key={scribble.id} className="tl-user-scribble" scribble={scribble} zoom={zoomLevel} />
+	))
 }
 
 function BrushWrapper() {
@@ -243,13 +234,9 @@ function SnapIndicatorWrapper() {
 
 	if (!(SnapIndicator && lines.length > 0)) return null
 
-	return (
-		<>
-			{lines.map((line) => (
-				<SnapIndicator key={line.id} className="tl-user-snapline" line={line} zoom={zoomLevel} />
-			))}
-		</>
-	)
+	return lines.map((line) => (
+		<SnapIndicator key={line.id} className="tl-user-snapline" line={line} zoom={zoomLevel} />
+	))
 }
 
 function HandlesWrapper() {
@@ -388,16 +375,12 @@ function ShapesWithSVGs() {
 		[editor]
 	)
 
-	return (
-		<>
-			{renderingShapes.map((result) => (
-				<Fragment key={result.id + '_fragment'}>
-					<Shape {...result} dprMultiple={dprMultiple} />
-					<DebugSvgCopy id={result.id} />
-				</Fragment>
-			))}
-		</>
-	)
+	return renderingShapes.map((result) => (
+		<Fragment key={result.id + '_fragment'}>
+			<Shape {...result} dprMultiple={dprMultiple} />
+			<DebugSvgCopy id={result.id} />
+		</Fragment>
+	))
 }
 function ReflowIfNeeded() {
 	const editor = useEditor()
@@ -448,70 +431,6 @@ function ShapesToDisplay() {
 	)
 }
 
-function ShapeIndicators() {
-	const editor = useEditor()
-	const renderingShapes = useValue('rendering shapes', () => editor.getRenderingShapes(), [editor])
-	const rPreviousSelectedShapeIds = useRef<Set<TLShapeId>>(new Set())
-	const idsToDisplay = useValue(
-		'should display selected ids',
-		() => {
-			// todo: move to tldraw selected ids wrappe
-			const prev = rPreviousSelectedShapeIds.current
-			const next = new Set<TLShapeId>()
-			if (
-				editor.isInAny(
-					'select.idle',
-					'select.brushing',
-					'select.scribble_brushing',
-					'select.editing_shape',
-					'select.pointing_shape',
-					'select.pointing_selection',
-					'select.pointing_handle'
-				) &&
-				!editor.getInstanceState().isChangingStyle
-			) {
-				const selected = editor.getSelectedShapeIds()
-				for (const id of selected) {
-					next.add(id)
-				}
-				if (editor.isInAny('select.idle', 'select.editing_shape')) {
-					const instanceState = editor.getInstanceState()
-					if (instanceState.isHoveringCanvas && !instanceState.isCoarsePointer) {
-						const hovered = editor.getHoveredShapeId()
-						if (hovered) next.add(hovered)
-					}
-				}
-			}
-
-			if (prev.size !== next.size) {
-				rPreviousSelectedShapeIds.current = next
-				return next
-			}
-
-			for (const id of next) {
-				if (!prev.has(id)) {
-					rPreviousSelectedShapeIds.current = next
-					return next
-				}
-			}
-
-			return prev
-		},
-		[editor]
-	)
-
-	const { ShapeIndicator } = useEditorComponents()
-	if (!ShapeIndicator) return null
-
-	return (
-		<>
-			{renderingShapes.map(({ id }) => (
-				<ShapeIndicator key={id + '_indicator'} shapeId={id} hidden={!idsToDisplay.has(id)} />
-			))}
-		</>
-	)
-}
-
 function HintedShapeIndicator() {
 	const editor = useEditor()
 	const { ShapeIndicator } = useEditorComponents()
@@ -521,13 +440,9 @@ function HintedShapeIndicator() {
 	if (!ids.length) return null
 	if (!ShapeIndicator) return null
 
-	return (
-		<>
-			{ids.map((id) => (
-				<ShapeIndicator className="tl-user-indicator__hint" shapeId={id} key={id + '_hinting'} />
-			))}
-		</>
-	)
+	return ids.map((id) => (
+		<ShapeIndicator className="tl-user-indicator__hint" shapeId={id} key={id + '_hinting'} />
+	))
 }
 
 function CursorDef() {
